+++ crates/ryo3-reqwest/src/lib.rs
@@ -1,5 +1,6 @@
 #![doc = include_str!("../README.md")]
 mod client;
+mod cookie;
 mod default_client;
 mod errors;
 mod fetch;
@@ -11,12 +12,14 @@ mod response_stream;
 mod user_agent;
 
 pub use client::RyHttpClient;
+pub use cookie::PyCookie;
 pub use errors::RyReqwestError;
 use pyo3::prelude::*;
 pub use response_parking_lot::RyResponse;
 pub use response_stream::RyResponseStream;
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<PyCookie>()?;
     m.add_class::<RyHttpClient>()?;
     m.add_class::<RyResponse>()?;
     m.add_class::<RyReqwestError>()?;